cleanup SongPlayer shortcut code a bit, create reusable hook
