+++ components/SongPlayer/SongPlayer.tsx
@@ -1,5 +1,5 @@
 import { Box, Typography, useMediaQuery, useTheme } from '@mui/material';
-import React, { useImperativeHandle } from 'react';
+import React, { useCallback, useImperativeHandle } from 'react';
 import { useRef } from 'react';
 
 import { Song, SourceData } from '@models';
@@ -65,27 +65,42 @@ const SongPlayer = React.forwardRef<SongPlayerHandle, Props>(
     const narrowViewport = useMediaQuery(theme.breakpoints.down('md'));
     const controlsContainerRef = useRef<HTMLDivElement>(null);
 
+    const handleKeyDown = useCallback((ev: KeyboardEvent) => {
+      if (ev.target === controlsContainerRef.current) {
+        // event is actually one we dispatched ourselves (see next if)
+        // event is apparently not really dispatched only on child?
+        // instead, it gets triggered from body downwards again?
+        return;
+      }
+      // try to forward keydown event to controls
+      if (controlsContainerRef.current) {
+        copyAndDispatchKeyboardEvent(ev, controlsContainerRef.current);
+      }
+      if (ev.key === ' ') {
+        // in any case, prevent default behavior of hitting space key ( == scrolling page)
+        ev.preventDefault();
+      }
+    }, []);
+
+    // imperative handle allows parent to forward keyboard events and let the player handle all of them
     useImperativeHandle(ref, () => ({
-      handleKeyDown: ev => {
-        if (ev.target === controlsContainerRef.current) {
-          // event is actually one we dispatched ourselves (see next if)
-          // event is apparently not really dispatched only on child?
-          // instead, it gets triggered from body downwards again?
-          return;
-        }
-        // try to forward keydown event to controls
-        if (controlsContainerRef.current) {
-          copyAndDispatchKeyboardEvent(ev, controlsContainerRef.current);
-        }
-        if (ev.key === ' ') {
-          // in any case, prevent default behavior of hitting space key ( == scrolling page)
-          ev.preventDefault();
-        }
-      },
+      handleKeyDown,
     }));
 
     return (
-      <Box>
+      <Box
+        // onKeyDownCapture keyDown handling is merely a fallback if player is not given more control
+        // over keyboard events from parent (via SongPlayerHandle's handleKeyDown())
+        onKeyDownCapture={(ev: React.KeyboardEvent) =>
+          handleKeyDown(ev.nativeEvent)
+        }
+        tabIndex={-1} // unfortunately, we have to set this for handler to work: https://stackoverflow.com/a/44434971/13727176
+        sx={{
+          ':focus': {
+            outline: 'none', // remove outline added due to tabIndex
+          },
+        }}
+      >
         <Box
           sx={{
             display: 'flex',
